requirements.txt
streamlit_chat
python-dotenv
langchain
openai